+++ packages/hardhat-rocketh/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hardhat-rocketh
 
+## 4.0.14
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.10
+
 ## 4.0.13
 
 ### Patch Changes