+++ packages/rocketh-deploy-proxy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocketh/deploy-proxy
 
+## 0.10.2
+
+### Patch Changes
+
+- unnamedAccounts
+- Updated dependencies
+  - @rocketh/deploy@0.10.2
+  - rocketh@0.10.10
+
 ## 0.10.1
 
 ### Patch Changes