+++ packages/rocketh-deploy-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/deploy-router
 
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/deploy@0.10.2
+  - rocketh@0.10.10
+
 ## 0.10.1
 
 ### Patch Changes