+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/deploy
 
+## 0.10.2
+
+### Patch Changes
+
+- unnamedAccounts
+- Updated dependencies
+  - rocketh@0.10.10
+
 ## 0.10.1
 
 ### Patch Changes